requirements.txt
fastapi==0.115.0
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.10.3
SQLAlchemy==2.0.43
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0.2
python-dotenv==1.0.0
email-validator==2.1.0
psycopg2-binary==2.9.7